backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
pillow>=10.0.0
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
ffmpeg-python>=0.2.0
numba==0.58.1
opencv-python>=4.8.0
mediapipe>=0.10.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation>=0.40b0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=2.0.0

# Optional: Uncomment when ready to integrate
# anthropic>=0.8.1
# face-recognition>=1.3.0
